scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.11
supabase==2.9.1
httpx==0.27.2
websockets==13.1
python-dotenv==1.0.0
